Vendor nginx-plus-go-sdk client
* Change MaxFails datatype in IC. `int64`->`int` to match the type specified in the SDK
* Update Convert, Config and Configurator to use vendored sdk
* Update unit tests to pass

+++ internal/nginx/nginx.go
@@ -57,7 +57,7 @@ type Upstream struct {
 type UpstreamServer struct {
 	Address     string
 	Port        string
-	MaxFails    int64
+	MaxFails    int
 	FailTimeout string
 	SlowStart   string
 }
